feat(cartes): Fabriquer et gérer les tuiles de relief depuis le menu
Fabrique C++ (WMS BIL 32 bits, pas par axe emboîté dans la maille,
reprise croisée avec fetch_relief.py), colonne Relief, touche L,
suppression, annonce avec volume et destination.

Colonne Résolution suffixée 2D / 3D / 3D (à refaire), un jeu à un
autre pas étant détecté à l'inventaire. Rouet d'attente animé pendant
les mesures, et ré-inventaire limité à la seule carte remaniée.

Bornes géographiques de la fabrique en double : en float, l'arrondi
décalait le relief jusqu'à 80 cm et l'orthophoto jusqu'à 0,3 m.

+++ src/app/cartes/fabrique/FabriqueBoucle.cpp
@@ -105,10 +105,8 @@ void Fabrique::boucle(std::filesystem::path dossierCarte,
         return;
     }
 
-    const double degParMLon = static_cast<double>(carte.lonMax - carte.lonMin) /
-                              static_cast<double>(carte.largeurM);
-    const double degParMLat = static_cast<double>(carte.latMax - carte.latMin) /
-                              static_cast<double>(carte.hauteurM);
+    const double degParMLon = (carte.lonMax - carte.lonMin) / static_cast<double>(carte.largeurM);
+    const double degParMLat = (carte.latMax - carte.latMin) / static_cast<double>(carte.hauteurM);
     const double tuileM = static_cast<double>(TUILE_PX) * static_cast<double>(mParPixel);
 
     const auto debut = std::chrono::steady_clock::now();
@@ -128,11 +126,10 @@ void Fabrique::boucle(std::filesystem::path dossierCarte,
 
             /* Emprise du bloc. La grille est ancrée sur le coin nord-ouest, et la
                rangée 0 est au nord : la latitude décroît quand la rangée croît. */
-            const double lonLo = static_cast<double>(carte.lonMin) +
-                                 static_cast<double>(col0) * tuileM * degParMLon;
+            const double lonLo = carte.lonMin + static_cast<double>(col0) * tuileM * degParMLon;
             const double lonHi = lonLo + static_cast<double>(nCol) * tuileM * degParMLon;
-            const double latHi = static_cast<double>(carte.latMax) -
-                                 static_cast<double>(rangee0) * tuileM * degParMLat;
+            const double latHi =
+                carte.latMax - static_cast<double>(rangee0) * tuileM * degParMLat;
             const double latLo = latHi - static_cast<double>(nRangee) * tuileM * degParMLat;
 
             bool recu = false;
